Add copyrights to all source, show image when launching with docker, move trtllm config parser into trtllm folder
